refactor: renamed jobListing routes and controllers to jobListings for consistency
